Refactored to put main tab in its own class
